client: Move client package import path
Move the client package out from the server package.
